fix: rename case sensitive file
